% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{annotateInter,linkSet-method}
\alias{annotateInter,linkSet-method}
\alias{annotateInter}
\title{Annotate linkSet with inter/intra chromosome interactions}
\usage{
\S4method{annotateInter}{linkSet}(x)
}
\arguments{
\item{x}{A linkSet object}
}
\value{
A linkSet object with an additional metadata column 'inter_type'
}
\description{
Annotate linkSet with inter/intra chromosome interactions
}
\examples{
data(linkExample)
linkExample <- annotateInter(linkExample)
}
