% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatConverter.R
\name{as.data.frame,linkSet-method}
\alias{as.data.frame,linkSet-method}
\title{coerce linkSet to DataFrame}
\usage{
\S4method{as.data.frame}{linkSet}(x)
}
\arguments{
\item{x}{A linkSet object}
}
\value{
A DataFrame object
}
\description{
coerce linkSet to DataFrame
}
