% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatConverter.R
\name{as.GInteractions,linkSet-method}
\alias{as.GInteractions,linkSet-method}
\title{export linkset to GInteractions}
\usage{
\S4method{as.GInteractions}{linkSet}(x)
}
\arguments{
\item{x}{A linkset object}
}
\value{
A GInteractions object
}
\description{
export linkset to GInteractions
}
\examples{
data(linkExample)
gi <- as.GInteractions(linkExample)
gi
}
