% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/formatConverter.R
\name{as.GInteractions}
\alias{as.GInteractions}
\alias{as.GInteractions,linkSet-method}
\title{Convert to GInteractions}
\usage{
as.GInteractions(x)

\S4method{as.GInteractions}{linkSet}(x)
}
\arguments{
\item{x}{A linkset object}
}
\value{
A GInteractions object
}
\description{
Convert linkSet object to GInteractions
}
\examples{
data(linkExample)
gi <- as.GInteractions(linkExample)
gi
}
