% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{character_Or_missing-class}
\alias{character_Or_missing-class}
\title{Character or Missing Class Union}
\description{
A class union of character vectors and missing values used in linkSet
package for optional character arguments.
}
\details{
This class is used internally by the linkSet package to handle optional
character arguments, particularly in the linkSet constructor and methods.
}
