% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statical.R
\name{checkSplitDataNumericalFit}
\alias{checkSplitDataNumericalFit}
\title{check.split.data.numerical.fit}
\usage{
checkSplitDataNumericalFit(split.data)
}
\arguments{
\item{split.data}{Split data for model fitting}
}
\value{
Logical indicating if the model can be fit

Check split data numerical fit

None, stops with error if data is invalid
}
\description{
Helper function to check if the chicane model can be fit on each element of a split data list.

Check if split data has valid numerical fit for model fitting
}
\keyword{internal}
