% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{cleanUnusedRegions}
\alias{cleanUnusedRegions}
\alias{clean_unused_regions}
\alias{cleanUnusedRegions,linkSet-method}
\alias{clean_unused_regions,linkSet-method}
\title{Clean Unused Regions}
\usage{
cleanUnusedRegions(x)

clean_unused_regions(x)

\S4method{cleanUnusedRegions}{linkSet}(x)

\S4method{clean_unused_regions}{linkSet}(x)
}
\arguments{
\item{x}{A linkSet object}
}
\value{
A linkSet object with unused regions removed
}
\description{
This function removes unused regions from a linkSet object to minimize memory usage.
}
\examples{
data(linkExample)
linkExample <- cleanUnusedRegions(linkExample)
}
