% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatConverter.R
\name{convertToGrange}
\alias{convertToGrange}
\title{Convert string intervals to GRanges}
\usage{
convertToGrange(intervals)
}
\arguments{
\item{intervals}{Interval data to convert}
}
\value{
GRanges object
}
\description{
Convert various interval formats to GRanges objects
}
\keyword{Convert}
\keyword{GRanges}
\keyword{internal}
\keyword{intervals}
\keyword{to}
