% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{countInteractibility,linkSet-method}
\alias{countInteractibility,linkSet-method}
\alias{countBaitOe}
\title{Count bait and oe}
\usage{
\S4method{countInteractibility}{linkSet}(x, baitRegions = TRUE)
}
\arguments{
\item{x}{A linkSet object}

\item{baitRegions}{Whether to count bait regions}
}
\value{
A linkSet object with counts for each unique interaction
}
\description{
This function calculate the number of trans interactions for each bait and oe. The word "interactibility" can refer to https://doi.org/10.1038\%2Fnature11279.
}
\examples{
data(linkExample)
linkSet = c(linkExample,linkExample)
linkSet = countInteractions(linkSet)
linkSet = countInteractibility(linkSet)
}
