% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{countInteractions,linkSet-method}
\alias{countInteractions,linkSet-method}
\alias{countBaitOe}
\title{Count Bait and Other End Interactions}
\usage{
\S4method{countInteractions}{linkSet}(x, baitRegions = TRUE)
}
\arguments{
\item{x}{A linkSet object}

\item{baitRegions}{Whether to count bait regions}
}
\value{
A linkSet object with counts for each unique interaction
}
\description{
This function takes a linkSet object and counts the number of interactions for each bait and other end.
}
\examples{
data(linkExample)
linkSet = c(linkExample,linkExample)
linkSet = countInteractions(linkSet)
linkSet
}
