% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper.R
\name{createSampleLinkSet}
\alias{createSampleLinkSet}
\title{Create sample linkSet object}
\usage{
createSampleLinkSet()
}
\value{
A linkSet object with sample data
}
\description{
Create a sample linkSet object for testing purposes
}
\keyword{internal}
