% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{crossGeneEnhancer,linkSet-method}
\alias{crossGeneEnhancer,linkSet-method}
\title{Cross gene enhancer}
\usage{
\S4method{crossGeneEnhancer}{linkSet}(x, score_threshold = NULL)
}
\description{
Cross gene enhancer
}
