% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{diagnoseLinkSet,linkSet-method}
\alias{diagnoseLinkSet,linkSet-method}
\alias{diagnoseLinkSet}
\title{Diagnose the linkSet object, return barplot of inter/intra interaction and distance distribution}
\usage{
\S4method{diagnoseLinkSet}{linkSet}(x)
}
\arguments{
\item{x}{A linkSet object}
}
\description{
Diagnose the linkSet object, return barplot of inter/intra interaction and distance distribution
}
\value{
A ggplot object containing two plots: 
1) A barplot showing the distribution of inter/intra chromosomal interactions 
2) A histogram showing the distance distribution of interactions
}
\examples{
data(linkExample)
diagnoseLinkSet(linkExample)
}
