% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statical.R
\name{.check.split.data.numerical.fit}
\alias{.check.split.data.numerical.fit}
\title{check.split.data.numerical.fit}
\usage{
.check.split.data.numerical.fit(split.data)
}
\arguments{
\item{split.data}{List of data.table objects with fragment interaction data}
}
\value{
Logical indicating if the model can be fit on all elements of the split data list
}
\description{
Helper function to check if the chicane model can be fit on each element of a split data list.
}
\keyword{internal}
