% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatConverter.R
\name{.convert_to_grange}
\alias{.convert_to_grange}
\title{Convert string intervals to GRanges}
\usage{
.convert_to_grange(intervals)
}
\description{
Convert string intervals to GRanges
}
\keyword{internal}
