% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statical.R
\name{.filter.fragments}
\alias{.filter.fragments}
\title{filter.fragments}
\usage{
.filter.fragments(
  linkSet,
  bait.filters = c(0, 1),
  target.filters = c(0, 1),
  verbose = FALSE
)
}
\arguments{
\item{linkSet}{LinkSet object containing interactions}
}
\value{
LinkSet object containing fragments that passed all filters
}
\description{
Filter low and high-interacting restriction fragments based on the total number of trans counts
}
