% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statical.R
\name{.verify.linkSet}
\alias{.verify.linkSet}
\title{verify.linkSet}
\usage{
.verify.linkSet(linkSet)
}
\arguments{
\item{linkSet}{Object to be verified.}
}
\value{
None
}
\description{
Verify that linkSet object is in expected format. Throws an error if object does not fit requirements.
}
