% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{enforceOrder}
\alias{enforceOrder}
\title{Enforce order of anchors}
\usage{
enforceOrder(anchor1, anchor2)
}
\arguments{
\item{anchor1}{First anchor indices}

\item{anchor2}{Second anchor indices}
}
\value{
List with ordered anchors
}
\description{
Ensure consistent ordering of anchor pairs
}
\keyword{internal}
