% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatConverter.R
\name{exportInterBed,linkSet-method}
\alias{exportInterBed,linkSet-method}
\title{Export linkSet to interBed format}
\usage{
\S4method{exportInterBed}{linkSet}(x, outfile)
}
\arguments{
\item{x}{A linkSet object}

\item{outfile}{Output file name}
}
\description{
Export linkSet to interBed format
}
