% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/formatConverter.R
\name{exportInterBed}
\alias{exportInterBed}
\alias{exportInterBed,linkSet-method}
\title{Export linkSet to interBed format}
\usage{
exportInterBed(x, outfile)

\S4method{exportInterBed}{linkSet}(x, outfile)
}
\arguments{
\item{x}{A linkSet object}

\item{outfile}{Output file path}
}
\value{
None. The function writes to the specified file.
}
\description{
Exports a linkSet object to a tab-delimited interBed format file
}
\details{
Export linkSet to interBed Format
}
\examples{
data(linkExample)
tmpfile <- tempfile(fileext = ".txt")
exportInterBed(linkExample, tmpfile)
cat(readLines(tmpfile), sep = "\n")
}
