% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatConverter.R
\name{exportToLinkSet}
\alias{exportToLinkSet}
\title{Export to linkSet format}
\usage{
exportToLinkSet(
  cd,
  scoreCol = "score",
  cutoff = 0,
  b2bcutoff = NULL,
  order = c("position", "score")[1],
  removeMT = TRUE
)
}
\description{
Export to linkSet format
}
\keyword{internal}
