% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatConverter.R
\name{exportWashU,linkSet-method}
\alias{exportWashU,linkSet-method}
\title{Export linkSet to WashU format}
\usage{
\S4method{exportWashU}{linkSet}(x, outfile)
}
\arguments{
\item{x}{A linkSet object}

\item{outfile}{Output file name}
}
\description{
Export linkSet to WashU format
}
