% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{extractDataFromLinkset}
\alias{extractDataFromLinkset}
\title{Extract data from linkSet for plotting}
\usage{
extractDataFromLinkset(linkset)
}
\arguments{
\item{linkset}{A linkSet object}
}
\value{
A data.frame with extracted data

Extracted data for visualization
}
\description{
Extract data from linkSet object for plotting
}
\keyword{internal}
