% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statical.R
\name{fit.model}
\alias{fit.model}
\title{fit.model}
\usage{
fit.model(
  linkSet,
  distance.bins = NULL,
  distribution = "negative-binomial",
  adjustment.terms = NULL,
  maxit = 100,
  epsilon = 1e-08,
  cores = 1,
  trace = FALSE,
  verbose = FALSE,
  interim.data.dir = NULL
)
}
\arguments{
\item{distribution}{Name of distribution of the counts. Options are 'negative-binomial',
'poisson', 'truncated-poisson', and 'truncated-negative-binomial'}

\item{adjustment.terms}{Character vector of extra terms to adjust for in the model fit.}

\item{maxit}{Maximum number of IWLS iterations for fitting the model (passed to \code{glm.control})}

\item{epsilon}{Positive convergence tolerance for Poisson and negative binomial models. Passed to \code{glm.control}}

\item{cores}{Integer value specifying how many cores to use to fit model for cis-interactions.}

\item{trace}{Logical indicating if output should be produced for each of model fitting procedure. Passed to \code{glm.control} or \code{gamlss.control}}

\item{verbose}{Logical indicating whether to print progress reports.}

\item{interim.data.dir}{Path to directory to store intermediate QC data and plots.}

\item{interaction.data}{data.table object containing interaction counts. Must contain columns distance, count, and bait_trans_count.}
}
\value{
Interactions data with expected number of interactions and p-values added.
}
\description{
Fit negative binomial model to obtain p-values for interactions.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Fit a negative binomial model for obtaining p-value for interactions. The data is first sorted by distance, and models
are fit separately in each quantile of the distance-sorted data.
}\if{html}{\out{</div>}}
}
\examples{
\donttest{
	data(bre80);
fit.model(bre80);
}

}
