% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{geom_linkset,linkSet-method}
\alias{geom_linkset,linkSet-method}
\title{Add Genome Links to Coverage Plot.}
\usage{
\S4method{geom_linkset}{linkSet}(
  linkSet,
  score.col = "count",
  score.threshold = NULL,
  score.color = c("grey70", "#56B1F7", "#132B43"),
  scale.range = 10,
  plot.space = 0.1,
  plot.height = 0.2,
  arrow.size = 0.05,
  remove_x_axis = FALSE,
  link_plot_on_top = FALSE,
  extend.base = 1e+06,
  show.rect = FALSE,
  x.range = NULL,
  log.scale = TRUE
)
}
\arguments{
\item{score.col}{Column index that contains score information, used when \code{file.type} is bedpe. Default: NULL.}

\item{score.threshold}{The score threshold, used when \code{score.col} is not NULL. Default: NULL.}

\item{score.color}{The score color vector. Default: c("grey70", "#56B1F7", "#132B43").}

\item{scale.range}{Scale the height of links according to width, should be greater than or equal to 1 (not scale). Default: 10.}

\item{plot.space}{Top and bottom margin. Default: 0.1.}

\item{plot.height}{The relative height of link to coverage plot. Default: 0.2.}

\item{arrow.size}{The size of the arrow head. Default: 0.02.}

\item{remove_x_axis}{Logical value, whether to remove the x-axis. Default: FALSE.}

\item{link_plot_on_top}{Logical value, whether to plot the link plot on top of the coverage plot. Default: FALSE.}

\item{extend.base}{Extend the base pair range to show more information. Default: 1000000.}

\item{show.rect}{Logical value, whether to add rect border to the plot. Default: FALSE.}

\item{x.range}{The range of x-axis to show. Default: NULL.}

\item{log.scale}{Logical value, whether to log1p the score. Default: TRUE.}

\item{link.file}{File contains region link information.}

\item{file.type}{The type of \code{link.file}, choose from bedpe, pairs. Default: bedpe.}

\item{plot.curve}{One of 'curve' or 'bezier', for the latter it is required to install package \code{ggforce}. Default: 'curve'.}
}
\value{
Plot.
}
\description{
Add Genome Links to Coverage Plot.
}
\references{
\url{https://stuartlab.org/signac/articles/cicero.html}
}
