% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{geomRange}
\alias{geomRange}
\title{Plot genomic ranges}
\usage{
geomRange(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Aesthetic mappings}

\item{data}{Data for plotting}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}
}
\value{
A ggplot2 layer that can be added to a plot.

ggplot layer for genomic ranges
}
\description{
Custom ggplot2 geom for plotting genomic ranges
}
\details{
\code{geomRange()} and \code{geom_half_range()} draw tiles that are designed to
represent range-based genomic features, such as exons. In combination with
\code{geom_intron()}, these geoms form the core components for visualizing
transcript structures.
}
\examples{
library(ggplot2)

# Create some example data
df <- data.frame(
  start = c(100, 200, 300, 400),
  end = c(150, 250, 350, 450),
  strand = c("+", "+", "-", "-"),
  gene = c("Gene1", "Gene1", "Gene2", "Gene2")
)

# Basic usage
ggplot(df) +
  geomRange(aes(xmin = start, xmax = end, y = gene))

Geom for plotting genomic ranges

}
\keyword{internal}
