% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{getDistOutput}
\alias{getDistOutput}
\title{Get distance output}
\usage{
getDistOutput(regs, ai1, ai2, type, inter_type)
}
\arguments{
\item{regs}{Genomic regions}

\item{ai1}{Anchor 1 indices}

\item{ai2}{Anchor 2 indices}

\item{type}{Distance type}

\item{inter_type}{Interaction type}
}
\value{
Distance calculations
}
\description{
Calculate distance metrics for genomic interactions
}
\keyword{internal}
