% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{linkSet,GRanges,GRanges,character_Or_missing-method}
\alias{linkSet,GRanges,GRanges,character_Or_missing-method}
\title{Create a linkSet object from input data}
\usage{
\S4method{linkSet}{GRanges,GRanges,character_Or_missing}(anchor1, anchor2, specificCol, metadata = list(), ...)
}
\arguments{
\item{anchor1}{For the first method, a character vector of bait names. For the second method, a GRanges object containing anchor1 regions.}

\item{anchor2}{A GRanges object containing anchor2 regions}

\item{specificCol}{Optional character vector specifying names for the baits. Can be either a column name from anchor1's metadata or a vector of names.}

\item{metadata}{Optional list of metadata to store}

\item{...}{Additional columns to add to the linkSet's elementMetadata}
}
\value{
A linkSet object containing the genomic interactions with metadata from both anchor regions
}
\description{
Create a linkSet object from input data
}
