% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{themeLinkset}
\alias{themeLinkset}
\alias{themeRange}
\title{linkSet-theme}
\usage{
themeLinkset(x.range, margin.len, show.rect)

themeRange(x.range, show.rect)
}
\arguments{
\item{x.range}{A numeric vector specifying the x-axis range.}

\item{margin.len}{A numeric value specifying the margin length.}

\item{show.rect}{A logical value indicating whether to show the rectangle.}
}
\description{
linkSet-theme
}
\examples{
data(linkExample)
x.range <- c(0, 100)
margin.len <- 0.5
show.rect <- TRUE
p <- ggplot2::ggplot() + 
  ggplot2::geom_blank() +
  themeLinkset(x.range, margin.len, show.rect) +
  ggplot2::scale_x_continuous(limits = x.range)
print(p)
data(linkExample)
x.range <- c(0, 100)
show.rect <- TRUE
p <- ggplot2::ggplot() + 
  ggplot2::geom_blank() +
  themeRange(x.range, show.rect) +
  ggplot2::scale_x_continuous(limits = x.range)
print(p)

}
