% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_genomic_ranges,linkSet-method}
\alias{plot_genomic_ranges,linkSet-method}
\alias{plot_genomic_ranges}
\title{plot genomic ranges and links}
\usage{
\S4method{plot_genomic_ranges}{linkSet}(
  linkset,
  showBait = NULL,
  showOE = NULL,
  x.range = NULL,
  score.col = "count",
  show.rect = TRUE,
  extend.base = 10000,
  ...,
  bait_col = "red",
  oe_col = "DeepSkyBlue3",
  default_col = "grey",
  vjust = NULL,
  linejoin = "mitre",
  na.rm = FALSE,
  minimal_width = 0.01,
  show.legend = NA,
  inherit.aes = TRUE,
  link_plot_on_top = FALSE,
  arrow.size = 0.05,
  remove_x_axis = FALSE,
  plot.height = 0.4,
  plot.space = 0.1,
  log.scale = TRUE
)
}
\arguments{
\item{linkset}{A \code{linkSet} object.}

\item{showBait}{A character vector specifying the bait region to be shown. Default: NULL.}

\item{showOE}{A \code{GRanges} object specifying the oe region to be shown. Default: NULL.}

\item{x.range}{A numeric vector of length 2 specifying the x-axis range. Default: NULL.}

\item{score.col}{A character string specifying the column name of the score. Default: "count".}

\item{show.rect}{Logical value, whether to show the rectangle. Default: TRUE.}

\item{extend.base}{A numeric value specifying the extension base. Default: 10000.}

\item{...}{Additional arguments.}

\item{bait_col}{A character string specifying the color of the bait region. Default: "red".}

\item{oe_col}{A character string specifying the color of the oe region. Default: "DeepSkyBlue3".}

\item{default_col}{A character string specifying the color of the default region. Default: "grey".}

\item{vjust}{A numeric value specifying the vertical justification. Default: NULL.}

\item{linejoin}{A character string specifying the line join. Default: "mitre".}

\item{na.rm}{Logical value, whether to remove NA values. Default: FALSE.}

\item{minimal_width}{A numeric value specifying the minimal width. Default: 0.01.}

\item{show.legend}{Logical value, whether to show the legend. Default: NA.}

\item{inherit.aes}{Logical value, whether to inherit the aesthetics. Default: TRUE.}

\item{link_plot_on_top}{Logical value, whether to plot the link plot on top of the coverage plot. Default: FALSE.}

\item{arrow.size}{A numeric value specifying the size of the arrow head. Default: 0.05.}

\item{remove_x_axis}{Logical value, whether to remove the x-axis. Default: FALSE.}

\item{plot.height}{A numeric value specifying the height of the plot. Default: 0.4.}

\item{plot.space}{A numeric value specifying the space between the plot and the link plot. Default: 0.1.}

\item{log.scale}{Logical value, whether to log scale the score. Default: TRUE.}
}
\value{
A \code{ggplot} object.
}
\description{
plot genomic ranges and links
}
\examples{
data(linkExample)
plot_genomic_ranges(linkExample, extend.base = 10)
}
