% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatConverter.R
\name{readvalidPairs}
\alias{readvalidPairs}
\title{Read validPairs file to GInteractions}
\usage{
readvalidPairs(file)
}
\arguments{
\item{file}{A character string specifying the path to the validPairs file}
}
\value{
A GInteractions object
}
\description{
Read validPairs file to GInteractions
}
