% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{themeLinkset}
\alias{themeLinkset}
\title{Theme for linkSet plots}
\usage{
themeLinkset(x.range, margin.len, show.rect)
}
\arguments{
\item{x.range}{The x-axis range}

\item{margin.len}{Margin length}

\item{show.rect}{Whether to show rectangle}
}
\value{
A ggplot2 theme
}
\description{
Theme for linkSet plots
}
