% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{themeRange}
\alias{themeRange}
\title{Theme for genomic range plots}
\usage{
themeRange(x.range, show.rect)
}
\arguments{
\item{x.range}{The x-axis range}

\item{show.rect}{Whether to show rectangle}
}
\value{
A ggplot2 theme
}
\description{
Theme for genomic range plots
}
