% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotatePromoter,linkSet-method}
\alias{annotatePromoter,linkSet-method}
\title{Annotate the link set with txDb. Give a gene list, and return a}
\usage{
\S4method{annotatePromoter}{linkSet}(
  x,
  genome = "hg38",
  keyType = "symbol",
  upstream = 500,
  overwrite = FALSE
)
}
\arguments{
\item{x}{linkSet}

\item{genome}{the genome you want to annotate}

\item{keyType}{the key type. You can check with AnnotationDbi::keytypes}

\item{upstream}{The upstream base from the gene}
}
\value{
linkSet object
}
\description{
Annotate the link set with txDb. Give a gene list, and return a
}
