% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{clean_unused_regions,linkSet-method}
\alias{clean_unused_regions,linkSet-method}
\title{clean unused regions}
\usage{
\S4method{clean_unused_regions}{linkSet}(x)
}
\description{
clean unused regions
}
\examples{
data(linkExample)
linkExample <- clean_unused_regions(linkExample)
}
