% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/formatConverter.R
\name{Convert}
\alias{Convert}
\alias{Convert,GInteractions-method}
\alias{Convert,data.frame-method}
\alias{Convert,Pairs-method}
\alias{Convert,ANY-method}
\title{Convert different object types to linkSet}
\usage{
Convert(x, ...)

\S4method{Convert}{GInteractions}(x, baitCol = NULL, ...)

\S4method{Convert}{data.frame}(x, source = "data.frame", baitCol = "gene", oeCol = "peak", ...)

\S4method{Convert}{Pairs}(x, baitCol = NULL, ...)

\S4method{Convert}{ANY}(x, ...)
}
\arguments{
\item{x}{An object of unsupported class}

\item{...}{Additional arguments (not used)}

\item{baitCol}{A character string specifying the column to use for bait naming}

\item{specificCol}{A character string specifying the column to use for bait naming}
}
\value{
A linkSet object

A linkSet object

A linkSet object

A linkSet object

Nothing, throws an error
}
\description{
Convert other data formats to linkSet. Currently supported: GInteractions, data.frame.
}
