% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{crossGeneEnhancer,linkSet-method}
\alias{crossGeneEnhancer,linkSet-method}
\alias{crossGeneEnhancer}
\title{Cross gene enhancer}
\usage{
\S4method{crossGeneEnhancer}{linkSet}(x, score_threshold = NULL)
}
\arguments{
\item{score_threshold}{The minimum score to filter interactions}
}
\value{
A linkSet object with filtered interactions
}
\description{
Cross gene enhancer
}
\examples{
linkSet = data(linkExample)
linkSet = c(linkSet, linkSet)
linkSet = countInteractions(linkSet)
linkSet = filterLinks(linkSet, filter_intra = FALSE, filter_unannotate = FALSE, distance = 100000)
linkSet = crossGeneEnhancer(linkSet, score_threshold = 10)
}
