% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{diagnoseLinkSet,linkSet-method}
\alias{diagnoseLinkSet,linkSet-method}
\alias{diagnoseLinkSet}
\title{Diagnose the linkSet object, return barplot of inter/intra interaction and distance distribution}
\usage{
\S4method{diagnoseLinkSet}{linkSet}(x)
}
\arguments{
\item{x}{A linkSet object}
}
\value{
Returns the input linkSet object with additional metadata columns for inter/intra interaction types and distances. Also prints diagnostic plots showing distance distribution and inter/intra interaction proportions.
}
\description{
Diagnose the linkSet object, return barplot of inter/intra interaction and distance distribution
}
\examples{
data(linkExample)
diagnoseLinkSet(linkExample)
}
