% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{diagnoseLinkSet,linkSet-method}
\alias{diagnoseLinkSet,linkSet-method}
\title{Diagnose the linkSet object, return barplot of inter/intra interaction and distance distribution}
\usage{
\S4method{diagnoseLinkSet}{linkSet}(x)
}
\description{
Diagnose the linkSet object, return barplot of inter/intra interaction and distance distribution
}
