% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{filterLinks,linkSet-method}
\alias{filterLinks,linkSet-method}
\title{Filter links for further analysis}
\usage{
\S4method{filterLinks}{linkSet}(x, filter_intra = TRUE, filter_unannotate = TRUE, distance = NULL)
}
\description{
Filter links for further analysis
}
