% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linkExample}
\alias{linkExample}
\title{Example linkSet Object}
\format{
A linkSet object with example interactions
}
\usage{
data(linkExample)
}
\description{
A dataset containing example genomic interactions in linkSet format
}
\examples{
data(linkExample)
show(linkExample)
}
\keyword{datasets}
