% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{anchor1,linkSet-method}
\alias{anchor1,linkSet-method}
\alias{anchor1}
\alias{anchor2,linkSet-method}
\alias{anchor2}
\alias{regions,linkSet-method}
\alias{regions}
\alias{seqinfo,linkSet-method}
\alias{seqinfo}
\title{linkSet-accessors}
\usage{
\S4method{anchor1}{linkSet}(x)

\S4method{anchor2}{linkSet}(x)

\S4method{regions}{linkSet}(x)

\S4method{seqinfo}{linkSet}(x)
}
\arguments{
\item{x}{A linkSet object}
}
\value{
For the getters, values in various slots of x are returned,
while for the setters, the slots of x are modified accordingly – see Details.

A vector of the regions
}
\description{
Methods to get and set fields in an linkSet object.
}
\examples{
data(linkExample)
anchor1(linkExample)
data(linkExample)
anchor2(linkExample)
data(linkExample)
regions(linkExample)
}
\author{
Gilbert Han
}
