% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{anchor1,linkSet-method}
\alias{anchor1,linkSet-method}
\alias{anchor1}
\alias{anchor2,linkSet-method}
\alias{anchor2}
\alias{regions,linkSet-method}
\alias{regions}
\title{linkSet-accessors}
\usage{
\S4method{anchor1}{linkSet}(x)

\S4method{anchor2}{linkSet}(x)

\S4method{regions}{linkSet}(x)
}
\arguments{
\item{x}{A linkSet object}
}
\value{
A vector of the first anchor

A vector of the second anchor

A vector of the regions
}
\description{
Get the first anchor of a linkSet object

Get the second anchor of a linkSet object

Get the regions of a linkSet object
}
