% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkSet-package.R
\docType{package}
\name{linkSet-package}
\alias{linkSet-package}
\title{linkSet: Base Classes for Storing Genomic Link Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The goal of linkSet is to provide tools for working with gene-enhancer interactions, which is commonly seen in HiC, PC-HiC, and single-cell ATAC-seq data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GilbertHan1011/linkSet}
  \item \url{https://gilberthan1011.github.io/linkSet}
  \item \url{https://gilberthan1011.github.io/linkSet/}
  \item Report bugs at \url{https://github.com/GilbertHan1011/linkSet/issues/new}
}

}
\author{
\strong{Maintainer}: Gilbert Han \email{GilbertHan1011@gmail.com} (\href{https://orcid.org/0009-0000-0293-2467}{ORCID})

}
\keyword{internal}
