% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{subsetBait,linkSet-method}
\alias{subsetBait,linkSet-method}
\alias{subsetBait}
\alias{subsetBaitRegion,linkSet-method}
\alias{subsetBaitRegion}
\alias{subsetOE,linkSet-method}
\alias{subsetOE}
\title{Subset linkSet object based on bait names}
\usage{
\S4method{subsetBait}{linkSet}(x, subset)

\S4method{subsetBaitRegion}{linkSet}(x, subset)

\S4method{subsetOE}{linkSet}(x, subset)
}
\arguments{
\item{x}{A linkSet object}

\item{subset}{A GRanges object specifying the regions to keep}
}
\value{
A new linkSet object containing only the specified bait interactions

A new linkSet object containing only the interactions with bait regions overlapping the subset

A new linkSet object containing only the interactions with oe regions overlapping the subset
}
\description{
Subset linkSet object based on bait names

Subset linkSet object based on bait regions

Subset linkSet object based on other end (oe) regions
}
