% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{theme_linkset}
\alias{theme_linkset}
\alias{theme_range}
\title{linkSet-theme}
\usage{
theme_linkset(x.range, margin.len, show.rect)

theme_range(x.range, show.rect)
}
\description{
linkSet-theme
}
\examples{
data(linkExample)
x.range <- c(0, 100)
margin.len <- 0.5
show.rect <- TRUE
p <- ggplot2::ggplot() + theme_linkset(x.range, margin.len, show.rect)
print(p)
data(linkExample)
x.range <- c(0, 100)
show.rect <- TRUE
p <- ggplot2::ggplot() + theme_range(x.range, show.rect)
print(p)

}
