% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{orderLinks,linkSet-method}
\alias{orderLinks,linkSet-method}
\alias{orderLinks}
\title{Order linkSet by mcols}
\usage{
\S4method{orderLinks}{linkSet}(x, by = "count", decreasing = TRUE)
}
\arguments{
\item{by}{The column name to order by}

\item{decreasing}{Whether to sort in decreasing order}
}
\value{
A linkSet object with ordered interactions
}
\description{
Order linkSet by mcols
}
\examples{
linkSet = data(linkExample)
linkSet = c(linkSet,linkSet)
linkSet = countInteractions(linkSet)
linkSet = filterLinks(linkSet, filter_intra = FALSE, filter_unannotate = FALSE, distance = 100000)
linkSet = orderLinks(linkSet, by = "count", decreasing = TRUE)
}
