% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotBaits}
\alias{plotBaits}
\title{Plot Baits}
\usage{
plotBaits(
  linkset,
  scoreCol = "score",
  countCol = "count",
  n = 4,
  baits = NULL,
  plotBaitNames = TRUE,
  plevel1 = 5,
  plevel2 = 3,
  outfile = NULL,
  width = 20,
  height = 20,
  extend.base = 1e+06,
  bgCol = "black",
  lev2Col = "blue",
  lev1Col = "red",
  ...
)
}
\arguments{
\item{linkset}{A linkSet object}

\item{...}{Additional plotting parameters}
}
\value{
A ggplot object
}
\description{
Plot baits in a linkSet object
}
